requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   openai
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
distro==1.9.0
    # via openai
exceptiongroup==1.3.1
    # via anyio
faiss-cpu==1.13.2
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.12.0
    # via openai
numpy==2.2.6
    # via faiss-cpu
openai==2.15.0
    # via -r requirements.in
packaging==25.0
    # via faiss-cpu
pdf2image==1.17.0
    # via -r requirements.in
pillow==12.1.0
    # via pdf2image
pydantic==2.12.5
    # via openai
pydantic-core==2.41.5
    # via pydantic
pypdf==6.6.0
    # via -r requirements.in
python-dotenv==1.2.1
    # via -r requirements.in
regex==2026.1.15
    # via tiktoken
requests==2.32.5
    # via tiktoken
sniffio==1.3.1
    # via openai
tiktoken==0.12.0
    # via -r requirements.in
tqdm==4.67.1
    # via openai
typing-extensions==4.15.0
    # via
    #   anyio
    #   exceptiongroup
    #   openai
    #   pydantic
    #   pydantic-core
    #   pypdf
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via requests
